# spec/api.openapi.yaml
openapi: 3.1.0
info:
  title: '<service> Management API'
  version: 0.1.0
  summary: Authenticated REST API for orgs, identities, keys, tokens, and audit.
  description: |
    Management API for <service>. Authenticated, served from a <service>-controlled
    host (e.g., api.<service>.tld). Used by web UI, CLI, CI automation, and any
    third-party integration (Ansible collection, Terraform provider, GitHub Action).

    The OpenAPI spec is the source of truth. .NET 10 server stubs and clients
    (TS/Go/Python) are generated from it.

    For the public, unauthenticated key-publishing endpoints (WKD + `.keys` + `.gpg`)
    served from per-customer domains, see `well-known.openapi.yaml`.
  contact:
    name: '<service> team'
  license:
    name: TBD
servers:
  - url: https://api.example.com
    description: Hosted SaaS (placeholder)
  - url: http://localhost:8080
    description: Local dev
...
security:
  - bearerToken: []

# ----------------------------------------------------------------------------
# Paths
# ----------------------------------------------------------------------------
paths:
...
  /v1/auth/credentials/ssh:
    get:
      tags: [credentials]
      operationId: listSshCredentials
      summary: List the current identity's SSH credentials.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SshCredential' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [credentials]
      operationId: registerSshCredential
      summary: Register a new SSH public key as a login credential.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterSshCredentialRequest' }
      responses:
        '201':
          description: Credential registered.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SshCredential' }
        default: { $ref: '#/components/responses/Error' }
  /v1/auth/credentials/ssh/{credentialId}:
    parameters:
      - in: path
        name: credentialId
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [credentials]
      operationId: revokeSshCredential
      summary: Revoke an SSH credential.
      responses:
        '204': { description: Revoked }
        default: { $ref: '#/components/responses/Error' }
  /v1/auth/credentials/webauthn:
    get:
      tags: [credentials]
      operationId: listWebAuthnCredentials
      summary: List the current identity's WebAuthn credentials.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WebAuthnCredential' }
        default: { $ref: '#/components/responses/Error' }
  /v1/auth/credentials/webauthn/begin-registration:
    post:
      tags: [credentials]
      operationId: beginWebAuthnRegistration
      summary: Start registering a new WebAuthn credential.
      responses:
        '200':
          description: Registration options issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebAuthnOptions' }
        default: { $ref: '#/components/responses/Error' }
  /v1/auth/credentials/webauthn/complete-registration:
    post:
      tags: [credentials]
      operationId: completeWebAuthnRegistration
      summary: Submit the attestation and store the new WebAuthn credential.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebAuthnRegistrationCompleteRequest' }
      responses:
        '201':
          description: Credential registered.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebAuthnCredential' }
        default: { $ref: '#/components/responses/Error' }
  /v1/auth/credentials/webauthn/{credentialId}:
    parameters:
      - in: path
        name: credentialId
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [credentials]
      operationId: revokeWebAuthnCredential
      summary: Revoke a WebAuthn credential.
      responses:
        '204': { description: Revoked }
        default: { $ref: '#/components/responses/Error' }

  # --- Current session ---------------------------------------------------
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: API token (created via POST /v1/api-tokens) or session bearer from OAuth/OIDC login.
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://example.com/.well-known/openid-configuration
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: domain_not_verified }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    SshCredential:
      type: object
      required: [id, identityId, publicKey, fingerprint, createdAt]
      properties:
        id: { type: string, format: uuid }
        identityId: { type: string, format: uuid }
        publicKey:
          type: string
          description: Single authorized_keys-format line.
        fingerprint:
          type: string
          description: SHA256 fingerprint, prefixed with "SHA256:".
        algorithm:
          type: string
          example: ssh-ed25519
        comment:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
    RegisterSshCredentialRequest:
      type: object
      required: [publicKey]
      properties:
        publicKey: { type: string }
    WebAuthnCredential:
      type: object
      required: [id, identityId, credentialId, createdAt]
      properties:
        id: { type: string, format: uuid }
        identityId: { type: string, format: uuid }
        credentialId:
          type: string
          description: WebAuthn Credential ID (base64url-encoded).
        algorithm:
          type: integer
          description: COSE algorithm identifier.
        aaguid:
          type: string
          format: uuid
          nullable: true
        nickname:
          type: string
          nullable: true
          description: Optional human label (e.g., "Work YubiKey").
        signCount:
          type: integer
          format: int64
        createdAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }

    # --- Sign-up + auth requests ----------------------------------------
...
    WebAuthnRegistrationCompleteRequest:
      type: object
      required: [ceremonyId, attestation]
      properties:
        ceremonyId: { type: string }
        nickname:
          type: string
          nullable: true
        attestation:
          type: object
          additionalProperties: true
    WebAuthnOptions:
      type: object
      description: |
        Opaque JSON document containing either PublicKeyCredentialCreationOptions
        (registration) or PublicKeyCredentialRequestOptions (assertion), plus a
        ceremonyId the server uses to correlate the second leg.
      required: [ceremonyId, options]
      properties:
        ceremonyId: { type: string }
        options:
          type: object
          additionalProperties: true